fix: don't copy host settings when forking a container
Due to some unknown issues in podman 4.4, when a container fork is created and the host config is copied, this causes the podman network iptables rules to be removed when the forked container exits.

By ignoring the copying of the container host config, then the iptables rules seem to be preserved. Anyway, the forked container does not need to have a network to perform its task.